Move application stuff into own folder
