+++ ui/src/Layout.tsx
@@ -10,7 +10,7 @@ import ScrollUpButton from './component/ScrollUpButton';
 import SettingsDialog from './component/SettingsDialog';
 import SnackBarHandler from './snack/SnackBarHandler';
 import * as config from './config';
-import Applications from './pages/Applications';
+import Applications from './application/Applications';
 import Clients from './pages/Clients';
 import Login from './pages/Login';
 import Messages from './pages/Messages';
