+++ ui/src/application/AddApplicationDialog.tsx
