+++ ui/src/inject.tsx
@@ -4,7 +4,7 @@ import {SnackManager} from './snack/SnackManager';
 import {MessagesStore} from './stores/MessagesStore';
 import {CurrentUser} from './stores/CurrentUser';
 import {ClientStore} from './stores/ClientStore';
-import {AppStore} from './stores/AppStore';
+import {AppStore} from './application/AppStore';
 import {inject as mobxInject, Provider} from 'mobx-react';
 import {WebSocketStore} from './stores/WebSocketStore';
